All client architecture implemented.
