--- nextjs-blog/pages/api/client/getClients.js
@@ -1,18 +1,0 @@
-//import clientPromise from "../../lib/mongodb";
-import clientPromise from "../../../lib/mongodb";
-
-export default async (req, res) => {
-   try {
-       const client = await clientPromise;
-       const db = client.db("FinalProject");
-
-       const clients = await db
-           .collection("Client")
-           .find({})
-           .toArray();
-
-       res.json(clients);
-   } catch (e) {
-       console.error(e);
-   }
-};
