+++ nextjs-blog/pages/api/forms/newClientForm.js
@@ -2,7 +2,6 @@ import { updateClient } from "../../../src/components/client/client_service"
 // const Client = require("../../../src/components/client/client");
 
 export default async function handler(req, res) {
-    // Get data submitted in request's body.
     var body = req.body
 
     //map to object if necessary later
